express_test_fininsh_0704

+++ express/src/router/transaction.ts
@@ -1,6 +1,6 @@
 // transaction.ts
 import { Router, Request, Response } from "express";
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient, Prisma } from "@prisma/client";
 import {
   transactionValidationRules,
   validate,
@@ -26,6 +26,10 @@ transactionRouter.get("/", async (req: Request, res: Response) => {
 // 特定取引の取得
 transactionRouter.get("/:id", async (req: Request, res: Response) => {
   const id = parseInt(req.params.id);
+  if (isNaN(id)) {
+    return res.status(400).send("無効なIDです。");
+  }
+
   try {
     const transaction = await prisma.transaction.findUnique({ where: { id } });
     if (transaction) {
@@ -45,18 +49,18 @@ transactionRouter.get("/:id", async (req: Request, res: Response) => {
 // 新しい取引の作成
 transactionRouter.post(
   "/",
-  transactionValidationRules(),
+  transactionValidationRules("POST"),
   validate,
   async (req: Request, res: Response) => {
     const { date, amount, type, details, userId } = req.body;
     try {
       const newTransaction = await prisma.transaction.create({
         data: {
-          date: new Date(date).toISOString(), // DateオブジェクトをISO 8601形式の文字列に変換
+          date: new Date(date).toISOString(),
           amount,
           type,
           details,
-          User: { connect: { id: userId } }, // userIdをUserと関連付ける
+          User: { connect: { id: userId } },
         },
       });
       res.status(201).json(newTransaction);
@@ -70,27 +74,37 @@ transactionRouter.post(
   },
 );
 
-// 特定取引の更新
+// 特定の取引の更新
 transactionRouter.put(
   "/:id",
-  transactionValidationRules(),
+  transactionValidationRules("PUT"),
   validate,
   async (req: Request, res: Response) => {
     const id = parseInt(req.params.id);
     const { date, amount, type, details, userId } = req.body;
     try {
+      if (isNaN(id)) {
+        return res.status(400).send("無効なIDです。");
+      }
+
       const updatedTransaction = await prisma.transaction.update({
         where: { id },
         data: {
-          date: new Date(date).toISOString(), // DateオブジェクトをISO 8601形式のオブジェクトに変換
+          date: new Date(date).toISOString(),
           amount,
           type,
           details,
-          User: { connect: { id: userId } }, // userIdをUserと関連付ける
+          User: { connect: { id: userId } },
         },
       });
       res.json(updatedTransaction);
     } catch (error) {
+      if (
+        error instanceof Prisma.PrismaClientKnownRequestError &&
+        error.code === "P2025"
+      ) {
+        return res.status(404).send("取引が見つかりません。");
+      }
       console.error("Error updating transaction:", (error as Error).message);
       res.status(500).json({
         error: "取引の更新に失敗しました。",
@@ -100,14 +114,24 @@ transactionRouter.put(
   },
 );
 
-// 特定取引の削除
+// 特定の取引の削除
 transactionRouter.delete("/:id", async (req: Request, res: Response) => {
   const id = parseInt(req.params.id);
   try {
+    if (isNaN(id)) {
+      return res.status(400).send("無効なIDです。");
+    }
+
     await prisma.transaction.delete({ where: { id } });
     res.status(204).send();
-  } catch (error: unknown) {
-    console.error("Error deleting transaction:", (error as Error).message);
+  } catch (error) {
+    if (
+      error instanceof Prisma.PrismaClientKnownRequestError &&
+      error.code === "P2025"
+    ) {
+      return res.status(404).send("取引が見つかりません。");
+    }
+
     res.status(500).json({
       error: "取引の削除に失敗しました。",
       details: (error as Error).message,
